# apps/server/admin-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/admin/roles": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "角色创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleVo"
                }
              }
            }
          }
        },
        "summary": "创建角色",
        "tags": [
          "admin",
          "角色管理"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取角色列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleRoleVo"
                  }
                }
              }
            }
          }
        },
        "summary": "获取所有角色",
        "tags": [
          "admin",
          "角色管理"
        ]
      }
    },
    "/api/admin/roles/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取角色成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDetailVo"
                }
              }
            }
          }
        },
        "summary": "根据ID获取角色",
        "tags": [
          "admin",
          "角色管理"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleVo"
                }
              }
            }
          }
        },
        "summary": "更新角色",
        "tags": [
          "admin",
          "角色管理"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色删除成功"
          }
        },
        "summary": "删除角色",
        "tags": [
          "admin",
          "角色管理"
        ]
      }
    },
...
    "/api/admin/permissions": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "权限创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionVo"
                }
              }
            }
          }
        },
        "summary": "创建权限",
        "tags": [
          "admin",
          "权限管理"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取权限列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionVo"
                  }
                }
              }
            }
          }
        },
        "summary": "获取所有权限",
        "tags": [
          "admin",
          "权限管理"
        ]
      }
    },
    "/api/admin/permissions/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取权限成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionVo"
                }
              }
            }
          }
        },
        "summary": "根据ID获取权限",
        "tags": [
          "admin",
          "权限管理"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionVo"
                }
              }
            }
          }
        },
        "summary": "更新权限",
        "tags": [
          "admin",
          "权限管理"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限删除成功"
          }
        },
        "summary": "删除权限",
        "tags": [
          "admin",
          "权限管理"
        ]
      }
    },
...
    "/api/admin/role-init/initialize": {
      "post": {
        "operationId": "initializeRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "初始化成功"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "初始化系统角色和权限",
        "tags": [
          "admin/role-init",
          "角色初始化管理"
        ]
      }
    },
    "/api/admin/role-init/reset": {
      "post": {
        "operationId": "resetRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "重置成功"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "重置系统角色和权限（危险操作）",
        "tags": [
          "admin/role-init",
          "角色初始化管理"
        ]
      }
    },
    "/api/admin/role-init/overview": {
      "get": {
        "operationId": "getRoleOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取角色权限概览",
        "tags": [
          "admin/role-init",
          "角色初始化管理"
        ]
      }
    },
    "/api/admin/users": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserVo"
                }
              }
            }
          }
        },
        "summary": "创建用户",
        "tags": [
          "admin",
          "用户管理"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取用户列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDetailVo"
                  }
                }
              }
            }
          }
        },
        "summary": "获取所有用户",
        "tags": [
          "admin",
          "用户管理"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailVo"
                }
              }
            }
          }
        },
        "summary": "根据ID获取用户",
        "tags": [
          "admin",
          "用户管理"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserVo"
                }
              }
            }
          }
        },
        "summary": "更新用户",
        "tags": [
          "admin",
          "用户管理"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户删除成功"
          }
        },
        "summary": "删除用户",
        "tags": [
          "admin",
          "用户管理"
        ]
      }
    }
  },
  "info": {
    "title": "hs-knowledge-base-server",
    "description": "火山知识库平台API文档",
    "version": "0.0.1",
    "contact": {
      "name": "火山",
      "url": "https://github.com/hs-knowledge-base/hs-knowledge-base",
      "email": "2633057734@qq.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "请输入 JWT 格式的 Token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员"
          },
          "permissionIds": {
            "description": "权限ID列表",
            "example": [
              "permission-uuid-1",
              "permission-uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "description": "角色属性",
            "example": {
              "level": "high",
              "department": "IT"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "PermissionSummaryVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "权限ID"
          },
          "code": {
            "type": "string",
            "description": "权限编码"
          },
          "name": {
            "type": "string",
            "description": "权限名称"
          },
          "type": {
            "type": "string",
            "description": "权限类型"
          },
          "reason": {
            "type": "string",
            "description": "权限说明"
          },
          "displayName": {
            "type": "string",
            "description": "权限显示名称"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "type"
        ]
      },
      "RoleVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员"
          },
          "attributes": {
            "type": "object",
            "description": "角色属性",
            "example": {
              "level": "high",
              "department": "IT"
            }
          },
          "permissions": {
            "description": "权限列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionSummaryVo"
            }
          },
          "users": {
            "description": "用户列表",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "createdAtLocal": {
            "type": "string",
            "description": "创建时间（本地格式）",
            "example": "2024/1/1"
          },
          "permissionCount": {
            "type": "number",
            "description": "权限数量"
          },
          "userCount": {
            "type": "number",
            "description": "用户数量"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "createdAtLocal"
        ]
      },
      "SimpleRoleVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色ID"
          },
          "name": {
            "type": "string",
            "description": "角色名称"
          },
          "description": {
            "type": "string",
            "description": "角色描述"
          },
          "createdAtLocal": {
            "type": "string",
            "description": "创建时间（本地格式）",
            "example": "2024/1/1"
          },
          "permissionCount": {
            "type": "number",
            "description": "权限数量"
          },
          "userCount": {
            "type": "number",
            "description": "用户数量"
          }
        },
        "required": [
          "id",
          "name",
          "createdAtLocal"
        ]
      },
      "RoleDetailVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员"
          },
          "attributes": {
            "type": "object",
            "description": "角色属性",
            "example": {
              "level": "high",
              "department": "IT"
            }
          },
          "permissions": {
            "description": "权限列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionSummaryVo"
            }
          },
          "users": {
            "description": "用户列表",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "createdAtLocal": {
            "type": "string",
            "description": "创建时间（本地格式）",
            "example": "2024/1/1"
          },
          "permissionCount": {
            "type": "number",
            "description": "权限数量"
          },
          "userCount": {
            "type": "number",
            "description": "用户数量"
          },
          "permissionsByType": {
            "type": "object",
            "description": "角色拥有的权限按类型分组"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "createdAtLocal"
        ]
      },
...
      "UserVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "john@example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "example": true
          },
          "firstName": {
            "type": "string",
            "description": "名字",
            "example": "张"
          },
          "lastName": {
            "type": "string",
            "description": "姓氏",
            "example": "三"
          },
          "attributes": {
            "type": "object",
            "description": "用户属性",
            "example": {
              "department": "IT",
              "level": 3
            }
          },
          "roles": {
            "description": "用户角色列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleVo"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "fullName": {
            "type": "string",
            "description": "全名",
            "example": "张三"
          },
          "createdAtLocal": {
            "type": "string",
            "description": "创建时间（本地格式）",
            "example": "2024/1/1"
          },
          "roleCount": {
            "type": "number",
            "description": "角色数量"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "权限编码",
            "example": "system.user.view"
          },
          "name": {
            "type": "string",
            "description": "权限名称",
            "example": "查看用户"
          },
          "type": {
            "type": "string",
            "description": "权限类型",
            "example": "button",
            "enum": [
              "module",
              "menu",
              "button"
            ]
          },
          "description": {
            "type": "string",
            "description": "权限描述",
            "example": "允许查看用户详细信息"
          },
          "path": {
            "type": "string",
            "description": "前端路由路径",
            "example": "/system/user"
          },
          "icon": {
            "type": "string",
            "description": "图标",
            "example": "user-o"
          },
          "sort": {
            "type": "number",
            "description": "排序值",
            "example": 1
          },
          "parentId": {
            "type": "string",
            "description": "父权限ID"
          }
        },
        "required": [
          "code",
          "name",
          "type"
        ]
      },
      "PermissionVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "权限ID"
          },
          "code": {
            "type": "string",
            "description": "权限编码",
            "example": "system.user.view"
          },
          "name": {
            "type": "string",
            "description": "权限名称",
            "example": "查看用户"
          },
          "type": {
            "type": "string",
            "description": "权限类型",
            "example": "BUTTON"
          },
          "path": {
            "type": "string",
            "description": "路径",
            "example": "/system/user"
          },
          "icon": {
            "type": "string",
            "description": "图标",
            "example": "user-o"
          },
          "sort": {
            "type": "number",
            "description": "排序值",
            "example": 1
          },
          "parentId": {
            "type": "string",
            "description": "父权限ID"
          },
          "children": {
            "description": "子权限列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionVo"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "displayName": {
            "type": "string",
            "description": "权限显示名称"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "type",
          "sort",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "example": true
          },
          "firstName": {
            "type": "string",
            "description": "名字",
            "example": "张"
          },
          "lastName": {
            "type": "string",
            "description": "姓氏",
            "example": "三"
          },
          "roleIds": {
            "description": "角色ID列表",
            "example": [
              "role-uuid-1",
              "role-uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "description": "用户属性",
            "example": {
              "level": 3,
              "clearance": "high"
            }
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "UserDetailVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "john@example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "example": true
          },
          "firstName": {
            "type": "string",
            "description": "名字",
            "example": "张"
          },
          "lastName": {
            "type": "string",
            "description": "姓氏",
            "example": "三"
          },
          "attributes": {
            "type": "object",
            "description": "用户属性",
            "example": {
              "department": "IT",
              "level": 3
            }
          },
          "roles": {
            "description": "用户角色列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleVo"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "fullName": {
            "type": "string",
            "description": "全名",
            "example": "张三"
          },
          "createdAtLocal": {
            "type": "string",
            "description": "创建时间（本地格式）",
            "example": "2024/1/1"
          },
          "roleCount": {
            "type": "number",
            "description": "角色数量"
          },
          "allPermissions": {
            "description": "用户拥有的所有权限（通过角色）",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}